requirements.txt
# UE Toolkit 依赖项

# ========== 必需依赖 ==========

# UI框架
PyQt6>=6.4.0

# 图片处理（用于资产管理模块的缩略图生成）
Pillow>=9.0.0

# 进程管理（用于资产管理模块的自动清理功能和性能监控）
psutil>=5.9.0

# 拼音转换（用于资产管理模块的搜索功能）
pypinyin>=0.49.0

# ========== 可选依赖 ==========

# 配置热重载（BaseLogic的watch_changes功能）
# 如果不安装，配置热重载功能将自动禁用
watchdog>=2.0.0

# URL可达性检查（InputValidator的check_reachable功能）
# 如果不安装，URL可达性检查将自动跳过
requests>=2.28.0

# YAML配置文件支持（ConfigManager）
PyYAML>=6.0

# ========== v0.1 AI Deep Integration 依赖 ==========
# 
# ⚠️ 打包优化说明：
# 以下 AI 依赖仅用于开发环境的语义分析功能（IntentEngine、EnhancedMemoryManager）
# 当前这些功能未被实际使用，因此在打包时会被排除以减小体积（从 311MB → 50-80MB）
# 
# 影响：
# - ✅ Ollama 本地模型聊天 - 正常工作（依赖 httpx）
# - ✅ API Key 调用 - 正常工作（依赖 requests）
# - ✅ 工具调用系统 - 正常工作
# - ❌ 语义意图识别 - 不可用（未使用的功能）
# - ❌ FAISS 向量记忆 - 不可用（未使用的功能）
# 
# 如需在打包后使用这些功能，请修改 packaging/ue_toolkit.spec：
# 1. 从 excludes 列表移除相关库
# 2. 添加到 hiddenimports 列表
# 3. 恢复 collect_data_files 调用
# 
# 参考备份文件：packaging/ue_toolkit.spec.backup
# ====================================================

# NumPy 版本锁定（FAISS 需要 1.x）
numpy>=1.24.0,<2.0  # 开发环境保留，打包时排除

# 语义意图解析（IntentEngine）
sentence-transformers>=2.2.0  # 开发环境保留，打包时排除

# 向量存储和检索（FAISS - 主存储）
faiss-cpu>=1.7.4  # 开发环境保留，打包时排除

# GitHub 代码搜索（GitHubConnector）
PyGithub>=1.59.0

# Ollama 本地模型支持（OllamaLLMClient）
httpx>=0.24.0

# 注意：已移除 chromadb（Windows 不稳定，已替换为 FAISS）

# ========== 安全依赖 ==========

# 系统密钥环集成（安全存储 API 密钥）
keyring>=24.0.0

# 加密库（敏感数据加密）
cryptography>=41.0.0

# ========== 开发依赖 ==========

# 测试框架
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0

# 类型检查
mypy>=1.7.0

# 代码质量检查
flake8>=6.1.0

# 安全扫描
bandit>=1.7.5

# ========== 国内镜像源推荐 ==========

# 如果使用pip安装依赖时速度较慢，可以使用以下国内镜像源：
# 
# 清华大学镜像源：
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/
#
# 阿里云镜像源：
# pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/
#
# 中科大镜像源：
# pip install -r requirements.txt -i https://pypi.mirrors.ustc.edu.cn/simple/
#
# 豆瓣镜像源：
# pip install -r requirements.txt -i https://pypi.douban.com/simple/
#
# 永久配置pip使用国内镜像源（推荐）：
# pip config set global.index-url https://pypi.tuna.tsinghua.edu.cn/simple/